+++ src/consumer/component/dependencies/dependency-resolver/dependencies-versions-resolver.js
@@ -10,9 +10,6 @@ import type Consumer from '../../../../consumer/consumer';
 import type { PathLinux } from '../../../../utils/path';
 import getNodeModulesPathOfComponent from '../../../../utils/bit/component-node-modules-path';
 import Dependencies from '../dependencies';
-import { MANUALLY_ADD_DEPENDENCY, MANUALLY_REMOVE_DEPENDENCY } from '../../../../constants';
-
-const isValidVersion = ver => ver !== MANUALLY_ADD_DEPENDENCY && ver !== MANUALLY_REMOVE_DEPENDENCY;
 
 /**
  * The dependency version is determined by the following strategies by this order.
@@ -108,7 +105,7 @@ export default function updateDependenciesVersions(consumer: Consumer, component
     if (!packageObject || R.isEmpty(packageObject)) return null;
     const packageId = Object.keys(packageObject)[0];
     const version = packageObject[packageId];
-    if (!semver.valid(version)) return null; // it's probably a relative path to the component
+    if (!semver.valid(version) && !semver.validRange(version)) return null; // it's probably a relative path to the component
     const validVersion = version.replace(/[^0-9.]/g, ''); // allow only numbers and dots to get an exact version
     return componentId.changeVersion(validVersion);
   }
@@ -129,8 +126,7 @@ export default function updateDependenciesVersions(consumer: Consumer, component
   }
 
   function getIdFromConsumerBitConfig(componentId: BitId): ?BitId {
-    const allDependenciesOverrides = component.overrides.getAllDependenciesOverridesFromConsumer();
-    const dependencies = R.filter(isValidVersion, allDependenciesOverrides);
+    const dependencies = component.overrides.getComponentDependenciesWithVersionFromConsumer();
     if (R.isEmpty(dependencies)) return null;
     const dependency = Object.keys(dependencies).find(
       idStr => componentId.toStringWithoutVersion() === idStr || componentId.toStringWithoutScopeAndVersion() === idStr
@@ -140,8 +136,7 @@ export default function updateDependenciesVersions(consumer: Consumer, component
   }
 
   function getIdFromComponentBitConfig(componentId: BitId): ?BitId {
-    const allDependenciesOverrides = component.overrides.getAllDependenciesOverrides();
-    const dependencies = R.filter(isValidVersion, allDependenciesOverrides);
+    const dependencies = component.overrides.getComponentDependenciesWithVersion();
     if (R.isEmpty(dependencies)) return null;
     const dependency = Object.keys(dependencies).find(
       idStr => componentId.toStringWithoutVersion() === idStr || componentId.toStringWithoutScopeAndVersion() === idStr